Use postgres instead of mysql
2 reasons

1) I'm trying to install this on heroku, and the env var for databases when using postgres is DATABASE_URL, which matches with sqlx tool's expected env var
2) I want to later store GIS info in the database, and it looks like postgres is the most commonly used database for that, via https://postgis.net

+++ src/model/county.rs
@@ -3,7 +3,7 @@ use sqlx::FromRow;
 
 #[derive(FromRow, Deserialize, Serialize)]
 pub struct County {
-    id: u16,
+    id: i16,
     name: String,
-    state: u8,
+    state: i16,
 }
