+++ src/model/data.rs
@@ -3,7 +3,7 @@ use sqlx::FromRow;
 
 #[derive(FromRow, Deserialize, Serialize)]
 pub struct Data {
-    pub county_id: u16,
-    pub state_id: u8,
-    pub value: f32,
+    pub county_id: i16,
+    pub state_id: i16,
+    pub value: f64,
 }
