+++ src/model/map_visualization.rs
@@ -3,17 +3,17 @@ use sqlx::FromRow;
 
 #[derive(FromRow, Deserialize, Serialize)]
 pub struct MapVisualization {
-    pub id: u64,
-    pub dataset: u64,
-    pub map_type: u8,
-    pub legend_ticks: Option<u8>,
+    pub id: i32,
+    pub dataset: i32,
+    pub map_type: i16,
+    pub legend_ticks: Option<i16>,
     pub should_normalize: bool,
-    pub color_palette: u8,
+    pub color_palette: i16,
     pub reverse_scale: bool,
     pub invert_normalized: bool,
-    pub scale_type: u8,
-    pub formatter_type: u8,
-    pub decimals: u8,
-    pub legend_formatter_type: Option<u8>,
-    pub legend_decimals: Option<u8>,
+    pub scale_type: i16,
+    pub formatter_type: i16,
+    pub decimals: i16,
+    pub legend_formatter_type: Option<i16>,
+    pub legend_decimals: Option<i16>,
 }
